initial commit
This is the initial commit for a project that is meant to contain code
and modules for the pickle interpreter - extending the interpreter with
operating system specific code, or with code that does not belong in the
core pickle interpreter as it would bloat it up.

The initial commit contains a rough implementation of the build system
and module layout, and a small test driver. There is much more to do,
such as; integrating the referenced sub-modules, making sure the build
system automatically initializes the git repositories if they are
missing, and writing more glue code.