Added skeleton module system
A skeleton module system for integrating the modules under the 'mod/'
directory has been added. It is expected that this will change shape as
it becomes clear what does and does not work.

The 'pickle' module has also been updated to version 4.1.2 of the
interpreter, which brings with it a lot of changes that make the
interpreter commands behave like their counterparts in an actual TCL
interpreter.